Feature iep3 (#21)
* Constexpr is implicitly inline

* Initial generic KML parser

- Parses gx:Track elements

* Provide abstract "KML track" class

* Add AbstractKMLParser

- Iterates over basic KML document structure
- FIX: proper double -> float conversion (waypoints store their
coordinates as float)
- FIX: flightradar24.com parser, parse Point (not Track)

* Only import first <gx:Track>

* Initial Flight Recorder CSV import

* Update flightradar24 and FlightAware names

- Add URL links to changelog

* Aircraft autocompleter now matches middle words

* Use C++ standard int types

* Parse engine and primary flight control data

* Import secondary flight control data

* FIX: Show spoiler position in percent

* Rename label to "Spoiler position"

* Import gear handles and light

* Use QFlags instead of reimplemented flag operations

- For convenience

+++ src/SkyConnect/src/Connect/FS2020SimConnectPlugin/src/FS2020SimConnectPlugin.cpp
@@ -237,9 +237,9 @@ void FS2020SimConnectPlugin::onStopRecording() noexcept
         Waypoint departureWaypoint;
         PositionData position = userAircraft.getPositionConst().getFirst();
         departureWaypoint.identifier = Waypoint::CustomDepartureIdentifier;
-        departureWaypoint.latitude = position.latitude;
-        departureWaypoint.longitude = position.longitude;
-        departureWaypoint.altitude = position.altitude;
+        departureWaypoint.latitude = static_cast<float>(position.latitude);
+        departureWaypoint.longitude = static_cast<float>(position.longitude);
+        departureWaypoint.altitude = static_cast<float>(position.altitude);
         departureWaypoint.localTime = flight.getFlightConditionConst().startLocalTime;
         departureWaypoint.zuluTime = flight.getFlightConditionConst().startZuluTime;
         departureWaypoint.timestamp = 0;
@@ -248,9 +248,9 @@ void FS2020SimConnectPlugin::onStopRecording() noexcept
         Waypoint arrivalWaypoint;
         position = userAircraft.getPositionConst().getLast();
         arrivalWaypoint.identifier = Waypoint::CustomArrivalIdentifier;
-        arrivalWaypoint.latitude = position.latitude;
-        arrivalWaypoint.longitude = position.longitude;
-        arrivalWaypoint.altitude = position.altitude;
+        arrivalWaypoint.latitude = static_cast<float>(position.latitude);
+        arrivalWaypoint.longitude = static_cast<float>(position.longitude);
+        arrivalWaypoint.altitude = static_cast<float>(position.altitude);
         arrivalWaypoint.localTime = d->currentLocalDateTime;
         arrivalWaypoint.zuluTime = d->currentZuluDateTime;
         arrivalWaypoint.timestamp = qMax(getCurrentTimestamp(), departureWaypoint.timestamp + 1);
